add wallet selection component & states for connect wallet sidebar contents
