+++ src/components/wallet-sidebar/connect-wallet-sidebar.tsx
@@ -5,12 +5,8 @@ import { useState } from 'react';
 import { IoMdClose } from 'react-icons/io';
 import { IoWalletSharp } from 'react-icons/io5';
 
-import { apiFetch } from '@/lib/api-fetch';
-import { WalletManager, NautilusConnector } from '@/lib/wallets';
-import { ChallengeResponse, ErgoAuthRequest, ErgoAuthResponse } from '@/types';
-
 import { SheetClose } from '../ui/sheet';
-import Wallet from './wallet';
+import { WalletSelection } from './wallet-selection';
 
 const volkhov = Volkhov({
   subsets: ['latin'],
@@ -18,42 +14,7 @@ const volkhov = Volkhov({
 });
 
 const ConnectWalletSidebar = () => {
-  const [selected, setSelected] = useState<'nautilus' | 'ergopay'>('nautilus');
-
-  const handleConnectButtonOnClick = async () => {
-    const wallet = new WalletManager(new NautilusConnector());
-
-    // Connect wallet
-    const connected = await wallet.connect();
-    if (!connected) throw new Error('Wallet connection rejected');
-
-    // Get address
-    const address = await wallet.getAddress();
-
-    // Request challenge from backend
-    const challengeResponse: ChallengeResponse = await apiFetch('/auth/ergo/challenge', {
-      method: 'POST',
-      body: JSON.stringify({ address }),
-    });
-
-    // Sign challenge
-    const proof = await wallet.signMessage(address, challengeResponse.challenge);
-
-    // Authenticate
-    const body: ErgoAuthRequest = {
-      address,
-      challenge: challengeResponse.challenge,
-      proof,
-      captchaToken: 'test-token',
-    };
-
-    const authResponse: ErgoAuthResponse = await apiFetch('/auth/ergo/auth', {
-      method: 'POST',
-      body: JSON.stringify(body),
-    });
-
-    console.log(authResponse);
-  };
+  const [state] = useState<'selection' | 'login'>('selection');
 
   return (
     // container
@@ -74,35 +35,7 @@ const ConnectWalletSidebar = () => {
         <span className='text-[22px]'>Connect your wallet</span>
       </div>
 
-      {/* wallets */}
-      <div className='flex flex-col space-y-2'>
-        {/* Natilus */}
-        <Wallet
-          onClick={() => setSelected('nautilus')}
-          src='/icons/natilus-40x40.png'
-          alt='Natilus icon'
-          size={40}
-          name='Natilus'
-          selected={selected == 'nautilus'}
-        />
-        {/* Ergo Pay */}
-        <Wallet
-          onClick={() => setSelected('ergopay')}
-          alt='Ergo Pay icon'
-          size={40}
-          name='Ergo Pay'
-          selected={selected == 'ergopay'}
-        />
-      </div>
-
-      {/* connect button */}
-      <button
-        className='h-11 w-25 cursor-pointer rounded-xl border border-green-400 bg-green-700 text-[17px] tracking-wider
-          text-white shadow-[-2px_2px_6px_0_rgba(0,0,0)]/20 shadow-black hover:bg-green-900 dark:shadow-white'
-        onClick={handleConnectButtonOnClick}
-      >
-        Connect
-      </button>
+      {state === 'selection' && <WalletSelection />}
     </div>
   );
 };
